+++ pytest_tap/tests/test_plugin.py
@@ -19,7 +19,7 @@ def sample_test_file(testdir):
         def test_skipped():
             assert False
 
-        @pytest.mark.xfail
+        @pytest.mark.xfail(reason='a reason')
         def test_broken():
             assert False
     """
@@ -48,7 +48,7 @@ def test_stream(testdir, sample_test_file):
             "ok 1 test_stream.py::test_ok",
             "not ok 2 test_stream.py::test_not_ok",
             "ok 3 test_stream.py::test_skipped # SKIP some reason",
-            "ok 4 test_stream.py::test_broken # SKIP ",
+            "ok 4 test_stream.py::test_broken # TODO expected failure: a reason",
             "1..4",
         ]
     )
@@ -86,3 +86,35 @@ def test_outdir(testdir, sample_test_file):
     outdir = testdir.tmpdir.join("results")
     testresults = outdir.join("testresults.tap")
     assert testresults.check()
+
+
+def test_xfail_strict_function(testdir):
+    """An xfail with strict on will fail when it unexpectedly passes.
+
+    The xfail should look like an xfail by including the TODO directive.
+    """
+    testdir.makepyfile(
+        """
+        import pytest
+
+        @pytest.mark.xfail(reason='a reason')
+        def test_unexpected_pass():
+            assert True
+
+        @pytest.mark.xfail(reason='a reason', strict=True)
+        def test_broken():
+            assert True
+    """
+    )
+    result = testdir.runpytest_subprocess("--tap-stream")
+
+    result.stdout.fnmatch_lines(
+        [
+            (
+                "ok 1 test_xfail_strict_function.py::test_unexpected_pass "
+                "# TODO unexpected success: a reason"
+            ),
+            "not ok 2 test_xfail_strict_function.py::test_broken # TODO",
+            "# [XPASS(strict)] a reason",
+        ]
+    )